refactor: change api methods exports
