+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import ContactForm from '@/components/ContactForm/ContactForm'
 import SliderSection from '@/components/Slider/SliderSection'
 import UsefulLinks from '@/page-components/main/UsefulLinks/UsefulLinks'
 import MainHero from '@/page-components/main/MainHero'
-import { getNewBuildingsOnMain } from '@/globals/api/methods/getNewBuildingsOnMain'
+import { getNewBuildingsOnMain } from '@/globals/api'
 
 export default async function Home() {
   const newBuildingsOnMain = await getNewBuildingsOnMain()
