+++ src/components/ContactForm/ContactForm.tsx
@@ -6,8 +6,7 @@ import Input from '@/ui/Input'
 import Section from '@/layout/Section'
 import SubmitButton from '@/ui/buttons/SubmitButton'
 import FormMap from '@/components/ContactForm/FormMap'
-import { getAddresses } from '@/globals/api/methods/getAddresses'
-import { getSocials } from '@/globals/api/methods/getSocials'
+import { getAddresses, getSocials } from '@/globals/api'
 
 import mobileBavel from '@/assets/static/decorative-bavel-mobile.svg'
 import bavel from '@/assets/static/decorative-bavel.svg'
